requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.0
Flasgger==0.9.5
PyMySQL==1.0.2
Werkzeug==2.0.3